Renderer to separate directory
