+++ src/main.rs
@@ -11,6 +11,7 @@ use crate::noise::Perlin;
 use std::path::Path;
 use rayon::prelude::*;
 use std::sync::Arc;
+use crate::renderer::{RgbRenderer, Renderer};
 
 mod vec;
 mod ray;
@@ -22,6 +23,7 @@ mod aabb;
 mod bvh;
 mod texture;
 mod noise;
+mod renderer;
 
 fn main() {
     let nx = 512;
@@ -329,64 +331,4 @@ fn _get_cam(nx: u32, ny: u32, t_off: f32, t_span: f32, ttl: i32) -> Camera {
         t_off, t_span,
         ttl,
     )
-}
-
-trait Renderer {
-    fn color(&self, r: &Ray) -> V3;
-}
-
-struct RgbRenderer {
-    pub hittable: Box<dyn Hittable>
-}
-
-impl Renderer for RgbRenderer {
-    fn color(&self, r: &Ray) -> V3 {
-        match self.hittable.hit(&r, 0.0001, 99999.0) {
-            Some(hit) => {
-                let emitted = hit.material.emmit(&hit);
-                return match hit
-                    .material
-                    .scatter(r, &hit)
-                    .and_then(Ray::validate) {
-                    Some(scattered) => { emitted.0 + scattered.attenuation * self.color(&scattered) }
-                    None => emitted.0
-                };
-            }
-            None => {
-//                let unit_direction = r.direction.unit();
-//                let t: f64 = 0.5 * (unit_direction.y + 1.0);
-                return V3::new(0.05088, 0.05088, 0.05088);
-//                return (1.0 - t) * V3::ones() + t * V3::new(0.5, 0.7, 1.0);
-            }
-        };
-    }
-}
-
-struct TtlRenderer{
-    hittable: Box<dyn Hittable>,
-    ttl: i32
-}
-impl Renderer for TtlRenderer {
-    fn color(&self, r: &Ray) -> V3 {
-        match self.hittable.hit(&r, 0.0001, 99999.0) {
-            Some(hit) => {
-                return match hit
-                    .material
-                    .scatter(r, &hit)
-                    .and_then(Ray::validate) {
-                    Some(scattered) => {
-                        ttl_color(scattered.ttl, self.ttl) * self.color(&scattered)
-                    }
-                    None => ttl_color(r.ttl, self.ttl)
-                };
-            }
-            None => {
-                return ttl_color(r.ttl, self.ttl)
-            }
-        };
-    }
-}
-
-fn ttl_color(ray_ttl: i32, max_ttl:i32) -> V3{
-    (ray_ttl as f64 / max_ttl as f64)* V3::ones()
 }
